Move graph-database from graph-node to util
