+++ packages/graph-node/src/utils.ts
@@ -7,15 +7,13 @@ import path from 'path';
 import fs from 'fs-extra';
 import debug from 'debug';
 import yaml from 'js-yaml';
-import { DeepPartial, EntityTarget, InsertEvent, Repository, UpdateEvent, ValueTransformer } from 'typeorm';
-import { ColumnMetadata } from 'typeorm/metadata/ColumnMetadata';
 import assert from 'assert';
 import _ from 'lodash';
 
-import { GraphDecimal } from '@cerc-io/util';
+import { TypeId } from '@cerc-io/util';
 import { MappingKey, StorageLayout } from '@cerc-io/solidity-mapper';
 
-import { TypeId, EthereumValueKind, ValueKind } from './types';
+import { EthereumValueKind } from './types';
 
 const log = debug('vulcanize:utils');
 
@@ -538,186 +536,6 @@ export const getSubgraphConfig = async (subgraphPath: string): Promise<any> => {
   return config;
 };
 
-export const toEntityValue = async (instanceExports: any, entityInstance: any, data: any, field: ColumnMetadata, type: string) => {
-  const { __newString, Value } = instanceExports;
-  const { isArray, propertyName, isNullable } = field;
-
-  const entityKey = await __newString(propertyName);
-  const entityValuePtr = await entityInstance.get(entityKey);
-  const subgraphValue = Value.wrap(entityValuePtr);
-  const value = data[propertyName];
-
-  // Check if the entity property is nullable.
-  // No need to set the property if the value is null as well.
-  if (isNullable && value === null) {
-    return;
-  }
-
-  const entityValue = await formatEntityValue(instanceExports, subgraphValue, type, value, isArray);
-
-  return entityInstance.set(entityKey, entityValue);
-};
-
-export const fromEntityValue = async (instanceExports: any, entityInstance: any, key: string): Promise<any> => {
-  const { __newString } = instanceExports;
-  const entityKey = await __newString(key);
-  const entityValuePtr = await entityInstance.get(entityKey);
-
-  return parseEntityValue(instanceExports, entityValuePtr);
-};
-
-const parseEntityValue = async (instanceExports: any, valuePtr: number) => {
-  const {
-    __getString,
-    __getArray,
-    BigInt: ASBigInt,
-    Bytes,
-    BigDecimal,
-    Value
-  } = instanceExports;
-
-  const value = Value.wrap(valuePtr);
-  const kind = await value.kind;
-
-  switch (kind) {
-    case ValueKind.STRING: {
-      const stringValue = await value.toString();
-      return __getString(stringValue);
-    }
-
-    case ValueKind.BYTES: {
-      const bytesPtr = await value.toBytes();
-      const bytes = await Bytes.wrap(bytesPtr);
-      const bytesStringPtr = await bytes.toHexString();
-
-      return __getString(bytesStringPtr);
-    }
-
-    case ValueKind.BOOL: {
-      const bool = await value.toBoolean();
-
-      return Boolean(bool);
-    }
-
-    case ValueKind.INT: {
-      return value.toI32();
-    }
-
-    case ValueKind.BIGINT: {
-      const bigIntPtr = await value.toBigInt();
-      const bigInt = ASBigInt.wrap(bigIntPtr);
-      const bigIntStringPtr = await bigInt.toString();
-      const bigIntString = __getString(bigIntStringPtr);
-
-      return BigInt(bigIntString);
-    }
-
-    case ValueKind.BIGDECIMAL: {
-      const bigDecimalPtr = await value.toBigDecimal();
-      const bigDecimal = BigDecimal.wrap(bigDecimalPtr);
-      const bigDecimalStringPtr = await bigDecimal.toString();
-
-      return new GraphDecimal(__getString(bigDecimalStringPtr)).toFixed();
-    }
-
-    case ValueKind.ARRAY: {
-      const arrayPtr = await value.toArray();
-      const arr = await __getArray(arrayPtr);
-      const arrDataPromises = arr.map((arrValuePtr: any) => parseEntityValue(instanceExports, arrValuePtr));
-
-      return Promise.all(arrDataPromises);
-    }
-
-    case ValueKind.NULL: {
-      return null;
-    }
-
-    default:
-      throw new Error(`Unsupported value kind: ${kind}`);
-  }
-};
-
-const formatEntityValue = async (instanceExports: any, subgraphValue: any, type: string, value: any, isArray: boolean): Promise<any> => {
-  const { __newString, __newArray, BigInt: ASBigInt, Value, ByteArray, Bytes, BigDecimal, id_of_type: getIdOfType } = instanceExports;
-
-  if (isArray) {
-    const dataArrayPromises = value.map((el: any) => formatEntityValue(instanceExports, subgraphValue, type, el, false));
-    const dataArray = await Promise.all(dataArrayPromises);
-    const arrayStoreValueId = await getIdOfType(TypeId.ArrayStoreValue);
-    const valueArray = await __newArray(arrayStoreValueId, dataArray);
-
-    return Value.fromArray(valueArray);
-  }
-
-  switch (type) {
-    case 'ID':
-    case 'String': {
-      const entityValue = await __newString(value);
-
-      return Value.fromString(entityValue);
-    }
-
-    case 'Boolean': {
-      return Value.fromBoolean(value ? 1 : 0);
-    }
-
-    case 'Int': {
-      return Value.fromI32(value);
-    }
-
-    case 'BigInt': {
-      const valueStringPtr = await __newString(value.toString());
-      const bigInt = await ASBigInt.fromString(valueStringPtr);
-
-      return Value.fromBigInt(bigInt);
-    }
-
-    case 'BigDecimal': {
-      const valueStringPtr = await __newString(value.toString());
-      const bigDecimal = await BigDecimal.fromString(valueStringPtr);
-
-      return Value.fromBigDecimal(bigDecimal);
-    }
-
-    case 'Bytes': {
-      const entityValue = await __newString(value);
-      const byteArray = await ByteArray.fromHexString(entityValue);
-      const bytes = await Bytes.fromByteArray(byteArray);
-
-      return Value.fromBytes(bytes);
-    }
-
-    // Return default as string for enum or custom type.
-    default: {
-      const entityValue = await __newString(value);
-
-      return Value.fromString(entityValue);
-    }
-  }
-};
-
-export const resolveEntityFieldConflicts = (entity: any): any => {
-  if (entity) {
-    // Remove fields blockHash and blockNumber from the entity.
-    delete entity.blockHash;
-    delete entity.blockNumber;
-
-    // Rename _blockHash -> blockHash.
-    if ('_blockHash' in entity) {
-      entity.blockHash = entity._blockHash;
-      delete entity._blockHash;
-    }
-
-    // Rename _blockNumber -> blockNumber.
-    if ('_blockNumber' in entity) {
-      entity.blockNumber = entity._blockNumber;
-      delete entity._blockNumber;
-    }
-  }
-
-  return entity;
-};
-
 export const toJSONValue = async (instanceExports: any, value: any): Promise<any> => {
   const { CustomJSONValue, JSONValueTypedMap, __newString, __newArray, id_of_type: getIdOfType } = instanceExports;
 
@@ -803,96 +621,4 @@ const getEthereumType = (storageTypes: StorageLayout['types'], type: string, map
   }
 
   return utils.ParamType.from(label);
-};
-
-export const fromStateEntityValues = (
-  stateEntity: any,
-  propertyName: string,
-  relations: { [key: string]: any } = {},
-  transformer?: ValueTransformer | ValueTransformer[]
-): any => {
-  // Parse DB data value from state entity data.
-  if (relations) {
-    const relation = relations[propertyName];
-
-    if (relation) {
-      if (relation.isArray) {
-        return stateEntity[propertyName].map((relatedEntity: { id: string }) => relatedEntity.id);
-      } else {
-        return stateEntity[propertyName]?.id;
-      }
-    }
-  }
-
-  if (transformer) {
-    if (Array.isArray(transformer)) {
-      // Apply transformer in reverse order similar to when reading from DB.
-      return transformer.reduceRight((acc, elTransformer) => {
-        return elTransformer.from(acc);
-      }, stateEntity[propertyName]);
-    }
-
-    return transformer.from(stateEntity[propertyName]);
-  }
-
-  return stateEntity[propertyName];
-};
-
-export const afterEntityInsertOrUpdate = async<Entity> (
-  frothyEntityType: EntityTarget<Entity>,
-  entities: Set<any>,
-  event: InsertEvent<any> | UpdateEvent<any>,
-  entityToLatestEntityMap: Map<new () => any, new () => any> = new Map()
-): Promise<void> => {
-  const entity = event.entity;
-
-  // Return if the entity is being pruned
-  if (entity.isPruned) {
-    return;
-  }
-
-  // Insert the entity details in FrothyEntity table
-  if (entities.has(entity.constructor)) {
-    const frothyEntity = event.manager.create(
-      frothyEntityType,
-      {
-        ..._.pick(entity, ['id', 'blockHash', 'blockNumber']),
-        ...{ name: entity.constructor.name }
-      }
-    );
-
-    await event.manager.createQueryBuilder()
-      .insert()
-      .into(frothyEntityType)
-      .values(frothyEntity as any)
-      .orIgnore()
-      .execute();
-  }
-
-  // Get latest entity's type
-  const entityTarget = entityToLatestEntityMap.get(entity.constructor);
-
-  if (!entityTarget) {
-    return;
-  }
-
-  // Get latest entity's fields to be updated
-  const latestEntityRepo = event.manager.getRepository(entityTarget);
-  const fieldsToUpdate = latestEntityRepo.metadata.columns.map(column => column.databaseName).filter(val => val !== 'id');
-
-  // Create a latest entity instance and upsert in the db
-  const latestEntity = getLatestEntityFromEntity(latestEntityRepo, entity);
-  await event.manager.createQueryBuilder()
-    .insert()
-    .into(entityTarget)
-    .values(latestEntity)
-    .orUpdate(
-      { conflict_target: ['id'], overwrite: fieldsToUpdate }
-    )
-    .execute();
-};
-
-export const getLatestEntityFromEntity = <Entity> (latestEntityRepo: Repository<Entity>, entity: any): Entity => {
-  const latestEntityFields = latestEntityRepo.metadata.columns.map(column => column.propertyName);
-  return latestEntityRepo.create(_.pick(entity, latestEntityFields) as DeepPartial<Entity>);
 };
