+++ packages/util/src/index.ts
@@ -13,9 +13,12 @@ export * from './types';
 export * from './indexer';
 export * from './job-runner';
 export * from './state-helper';
-export * from './graph-decimal';
 export * from './index-block';
 export * from './metrics';
 export * from './gql-metrics';
 export * from './common';
 export * from './server';
+export * from './graph/graph-decimal';
+export * from './graph/database';
+export * from './graph/utils';
+export * from './graph/types';
