+++ packages/util/src/misc.ts
@@ -17,7 +17,7 @@ import { EthClient } from '@cerc-io/ipld-eth-client';
 import { DEFAULT_CONFIG_PATH } from './constants';
 import { GQLCacheConfig, Config } from './config';
 import { JobQueue } from './job-queue';
-import { GraphDecimal } from './graph-decimal';
+import { GraphDecimal } from './graph/graph-decimal';
 import * as EthDecoder from './eth';
 import { getCachedBlockSize } from './block-size-cache';
 import { ResultEvent } from './indexer';
